ath10k: work-around fwcfg of 10 rc objs for 9980/84.
CT 10.4 firmware recently gained support for configuring
the RC objects.  But, the fwcfg files have been using 10,
and only 7 can be supported when using 64 vdevs.  So,
add special driver hack to work around this.  Otherwise,
firmware fails to load and user-space will not even discover
the radios in order to fix the config.

Signed-off-by: Ben Greear <greearb@candelatech.com>