ath10k: support mac detection from of_net
Add support for mac detection from of_net using the generic
of_get_mac_address. This is useful if for example the mac is
stored in a nvmem cell and needs to be applied instead of a
random mac address in case device_get_mac_address fail to
find a valid mac.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>
Signed-off-by: Ben Greear <greearb@candelatech.com>